Use the linker script variables for MPU setup for Keil Simulator Demo
Earlier we were using hard-coded addresses for MPU setup which
were ensured to be the same as linker script setup. This change
updates the Keil Simulator demo to use the variables exported
from the linker script. This ensures that the MPU setup does not
go out of sync with linker script.

git-svn-id: https://svn.code.sf.net/p/freertos/code/trunk@2761 1d2547de-c912-0410-9cb9-b8ca96c0e9e2