Use linker script variables for MPU setup for Nuvoton M2351 Keil Project
Earlier we were using hard-coded addresses for MPU setup which
were ensured to be the same as linker script setup. This change
updates the Keil uVision project for Nuvoton Numaker-PFM-M2351
to use the variables exported from the linker script. This ensures
that the MPU setup never goes out of sync with linker script.

git-svn-id: https://svn.code.sf.net/p/freertos/code/trunk@2762 1d2547de-c912-0410-9cb9-b8ca96c0e9e2